cleaned index.ts
